Chore: Relocate requests into their own package
