+++ src/main/java/dev/alejandro/tennis_game/player/PlayerController.java
@@ -3,7 +3,6 @@ package dev.alejandro.tennis_game.player;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +11,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import dev.alejandro.tennis_game.player.requests.CreatePlayerRequest;
+import dev.alejandro.tennis_game.player.requests.UpdatePlayerRequest;
 
 @RestController
 @RequestMapping("api/player")
