--- src/main/java/dev/alejandro/tennis_game/player/PlayerRequest.java
@@ -1,6 +1,0 @@
-package dev.alejandro.tennis_game.player;
-
-public interface PlayerRequest {
-
-    Player toEntity();
-}
