+++ src/main/java/dev/alejandro/tennis_game/player/PlayerService.java
@@ -4,6 +4,9 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import dev.alejandro.tennis_game.player.requests.PlayerRequest;
+import dev.alejandro.tennis_game.player.requests.UpdatePlayerRequest;
+
 @Service
 public class PlayerService {
