+++ src/main/java/dev/alejandro/tennis_game/player/requests/PlayerRequest.java
@@ -1,0 +1,8 @@
+package dev.alejandro.tennis_game.player.requests;
+
+import dev.alejandro.tennis_game.player.Player;
+
+public interface PlayerRequest {
+
+    Player toEntity();
+}
